% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blastp_df}
\alias{blastp_df}
\title{BLASTP Results for test Proteins}
\format{
A data frame with multiple rows and 12 variables:
\describe{
\item{qaccver}{Character. Protein query identifier.}
\item{saccver}{Character. Subject protein identifier from reference databases.}
\item{pident}{Numeric. Percentage identity between query and subject sequences.}
\item{length}{Numeric. Length of the aligned sequence region.}
\item{mismatch}{Numeric. Number of mismatches in the alignment.}
\item{gapopen}{Numeric. Number of gap openings in the alignment.}
\item{qstart}{Numeric. Start position in query sequence.}
\item{qend}{Numeric. End position in query sequence.}
\item{sstart}{Numeric. Start position in subject sequence.}
\item{send}{Numeric. End position in subject sequence.}
\item{evalue}{Numeric. Expect value for the alignment.}
\item{bitscore}{Numeric. Bit score for the alignment.}
}
}
\usage{
blastp_df
}
\description{
A dataset containing BLASTP alignment results for proteins from test genomes of bacteria,
including alignment metrics.
}
\keyword{datasets}
