% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_locate.R
\name{orf_locate}
\alias{orf_locate}
\title{Parse ORF Coordinates from Prodigal FASTA Headers}
\usage{
orf_locate(in_seq_data = seq_data)
}
\arguments{
\item{in_seq_data}{A data frame with two columns:
\describe{
\item{\code{SeqName}}{ORF Sequence Name.} \code{SeqName} includes the information of \code{qaccver},
\code{start}, \code{end}, and \code{direction}.
\item{\code{Sequence}}{ORF Sequence.}. e.g., "Kuafubacteriaceae--GCA_016703535.1---JADJBV010000001.1_1
# 74 # 1018 # 1 # ID=1_1;partial=00;start_type=ATG;rbs_motif=AGGAG;rbs_spacer=5-10bp;gc_cont=0.685"
}.
This data frame can be scaled from \strong{Prodigal} FASTA output by the command:
seq_data = "Prodigal.fasta" \%>\%
{Biostrings::readBStringSet(., format="fasta", nrec=-1L, skip=0L, seek.first.rec=FALSE, use.names=TRUE)} \%>\%
{data.frame(Sequence = .)} \%>\%
{tibble::rownames_to_column(., "SeqName")}

\strong{Prodigal} FASTA output: Each row must correspond to a header line (starting with \code{>}).
The header is expected to follow the standard Prodigal format:
\code{>contig_id # start # end # strand # ...}.}
}
\value{
A data frame
}
\description{
Extracts ORF identifiers, start/end positions and strand
orientation directly from the FASTA headers produced by
Prodigal.  The resulting table is ready for downstream
gene-cluster analyses.
}
\examples{
\dontrun{
loc <- orf_locate(data)
}
}
